requirements.txt
coverage>=5.5
python-dotenv>=0.19.0
Flask>=2.0
flask-cors>=3.0.10
gunicorn>=20.1
pytest>=6.2
